Unfinished, explanation stage done
Unfinished, explanation stage done

The bot starts by asking the user how they feel on a scale from 1 to 10. Then it runs through an explanation of how the bot works.

The current stumbling block is the fact that jquery doesn't seem to work on dynamically generated content. So when we get to asking the user to enter content that they've written, the jquery is supposed to recognise the user hitting the enter key, but it doesn't.

Also the storing of data isn't done yet in this version (this has to be added to the Python file)