v1.02 some bot responses are split up
Some more thought may be needed about the amount of time that the responses wait before appearing.

Note: the approach to this is fairly "hard-coded", i.e. it can only accommodate a bot response split specifically into one chunk (as normal) or two chunks, but not more -- doing more than two chunks would require extra code. I tried to do this differently, using a for loop to cycle through all the relevant chunks/fragments of bot response, which would have been flexible enough to cover any number of chunks/fragments of bot response. However I could'nt get this to work. The for loop continued running in parallel alongside the setTimeout function, and then by the time the setTimeout function was activated, the for loop had already got to the end.