v1.04 (broken) stop button appears after intro
Because of the problems alluded to in the comments to v1.03, this version attempts to simply not have the stop button on the screen in the initial bit. However implementing this in the way done in this code doesn't work. It causes it to not recognise the main getBotResponse function, and give an Uncaught ReferenceError