v3 slight tweak to wording re stop button
In section 9 where the stop button is explained, the wording in version2 was a bit lengthy, so this has been updated and put live. This is in response to the fact that (although the introduction of the stop button was supposed to improve the response rate) it's actually got worse since the stop button was added. 

(We moved to version 2 with the stop button on Thurs 5th/Mar, now updating to the updated wording in version 3 on Sun 8th Mar)