v4 more sophisticated bot responses
Responses: instead of just using generic/random responses, this version has a number of tailored responses based on what the user has just said
Wording re stop button: Some users were pressing the stop button the moment it appeared. To avoid that confusion, have added in some wording to clarify that this is not the intended behaviour. (Could do something more fancy in a later release?)
write_data: previous implementations caused the data to stop recording the moment the user selected "confidential"; however this meant that we don't gather the data about when the user drops out -- CORRECTION! This change didn't happen in this release -- to be done in next release instead
Added some more encouraging noises to the list. Also removed the encouraging noise that sais something like "remember, this is like an online journal"
Improved the way the javascript decides how long to have the typing ellipsis on the screen (i.e. the bot thinking time)