v6 removing leadString and other changes
Logic changed in the checking section. Now it also removes a lead string, which means that if the user's message starts with "feeling low", the bot treats this as if the user's message included "I'm feeling low"
Added in thisBotIsBadArray
note that thisbotisbadarray has some non-stadnard logic in
Added in a new method for checking  "have anything to live for", which involves just seeing if those words are in the user's message, and in that order
Adjusted the short response logic (i.e. the response which starts with "I see you've said something short, which is cool :-)"
Included a switch called "frontEnd" which allows the code to switch between the old and new front end
updates to imWorriedResponseAlreadyUsed and related logic
Tidied up the imGoingToKillMyselfArray: realised that it already includes a reflexive pronoun, so need to include the leading "I'm"
Added more content to iHaveNoWayOutArray and hadEnoughOfLifeArray
Tidied up feelingLowDownTerribleArray and then added more content to it
Added in iHateHowIFeelArray
Added a bit more content for imSadArray
Added in a few more extraneous words for the extraneousWordsArray
Strenghthened cleanText function, previously it took anything apart from alphabetic and space characters, however that left in a circumflex, which is part of the jumble of characters that apostrophe gets turned into, so now it takes out a circumflex as well
Some fixes for iDontKnowWhatToSayArray functionality