v7 includes API for GT and new bootstrap frontend
Created an API which will talk to Guided Track
Updated the server file so it can recognise and talk to the original javascript file and the new bootstrap js, as well as the API
This is currently set up to randomly give either the old or the new frontend. This is for A/B testing purposes
write_data: previous implementations caused the data to stop recording the moment the user selected "confidential"; however this meant that we don't gather the data about when the user drops out. Now fixed
On a related note, it was also failing to gather data about the final happiness score and the final feedback, so this has been changed so it will always be stored, regardless of what confindentiality settings the user selected
Minor tweaks to wording of bot response for the msgSaysNothingToLiveFor scenario
Updated the code so that the frontend tells the server what the nature of the frontend is (i.e. it distinguishes between different types of frontend). This allows it to tailor the wording so that it can refer to the stop buttons corrrectly, for example.
Expanded the "I hate how I look" array a little
Only change to the original html/js file is an adjustment so that it sends a suitable clientId back to the server, so it knows which client it's talking to