v1.1 for bootstrap frontend
Initially went live with a version of this, and then after going live, made a few more tweaks (based discussions between Daniel Wyatt and Sanjay, mostly intended to better mobile-optimise). Then made some updates to the code on the back of that, which are reflceted here. As the changes happened on the same day, there were actually no more users using the earlier (v1) version of the code.