v9 changes to intro section
The part of the algorithm which determines what the bots says when it's in the free text section is unchanged.

The changes are in the initial section, where the intro has been updated to mirror what originally happened in Guided Track. I.e. here: https://www.guidedtrack.com/programs/zwnwjov/preview

Once the intro has happened, there's an explanation section which is essentially the same as what currently happens in v8 (or v7 or probably several previous versions)