v10 gathers google ads data + bug fixes
Main change was that the code has been updated to incorporate the possibility that google ads might include a query string providing info about, e.g. the location of the user or their device, or which ad campaign the user came from.

This has been set up to be robust enough that it will work even if the query string is not provided.

Also the query string will be wiped away immediately.

There were also some bugs in the way that the previous published version (v9) interacted with the Guided Track bot. Specifically:
- there were complications with the confidential option (fixed with a tweak to the guided track code)
- section number 3.5 didn't work because it didn't accept a section number which wasn't typed as an int
- some adjustments were needed (in both the GT code and on line 2832 of the server code) to ensure that the abort route worked (i.e. if the initial happiness score is too high)